requirements.txt
asgiref==3.5.2
boto3==1.24.49
botocore==1.27.49
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
